django-backend/requirements.txt
amqp==2.6.1
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
boto3==1.15.16
botocore==1.18.16
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.36
cryptography==3.1.1
decorator==4.4.2
Django==2.2.16
django-celery-beat==2.0.0
django-filter==2.4.0
django-cors-headers==3.7.0
django-jet2==1.0.12
django-js-asset==1.2.2
django-mptt==0.11.0
django-otp==1.0.0
django-timezone-field==4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-yasg2==1.19.4
fabric==2.5.0
GitPython==3.1.18
hvac==0.10.5
idna==2.10
importlib-metadata==3.1.1
inflection==0.5.1
invoke==1.4.1
ipdb==0.13.4
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==1.1.1
natsort==7.1.0
packaging==20.7
paramiko==2.7.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0